requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
fastmcp>=0.9.0
neo4j>=5.0.0
networkx>=3.2
httpx>=0.27.0
diskcache>=5.6.3
pandas>=2.0.0
pydantic>=2.0.0
streamlit>=1.35.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
memory-profiler>=0.61.0
jsonschema>=4.22.0
